jvhdklj vh dvrjh

+++ ResoflexClientHandlingSystem/ScheduleForms/ScheduleForm.cs
@@ -2,6 +2,7 @@
 using ResoflexClientHandlingSystem.Core;
 using ResoflexClientHandlingSystem.Role;
 using ResoflexClientHandlingSystem.ScheduleForms;
+using ResoflexClientHandlingSystem.ScheduleForms.Reports;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -30,6 +31,10 @@ namespace ResoflexClientHandlingSystem
             //autocomplete data source
             projectName.AutoCompleteCustomSource = projectNameAutoComplete();
             clientName.AutoCompleteCustomSource = clientNameAutoComplete();
+
+            //tiles initilization
+            totalScheduleTile();
+            incompleteScheduleTile();
         }
 
         private void ScheduleForm_Load(object sender, EventArgs e)
@@ -46,15 +51,20 @@ namespace ResoflexClientHandlingSystem
             schedule.ScheduleId = int.Parse(dr.Cells[0].Value.ToString());
             schedule.ProjectOfSchedule = new Project(int.Parse(dr.Cells[1].Value.ToString()));
 
-            if (Database.deleteSchedule(schedule))
-            {
-                MessageBox.Show("Schedule Successfully Deleted\n", "Schedule Deleted", MessageBoxButtons.OK);
+            DialogResult res = MessageBox.Show("Are you sure you want delete this schedule?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
 
-                scheduleGrid.DataSource = getSchedules();
-            }
-            else
+            if (res == DialogResult.Yes)
             {
-                MessageBox.Show("Something went wrong!\n", "Schedule Deleted", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                if (Database.deleteSchedule(schedule))
+                {
+                    MessageBox.Show("Schedule Successfully Deleted\n", "Schedule Deleted", MessageBoxButtons.OK);
+
+                    scheduleGrid.DataSource = getSchedules();
+                }
+                else
+                {
+                    MessageBox.Show("Something went wrong!\n", "Schedule Deleted", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
             }
         }
 
@@ -75,8 +85,9 @@ namespace ResoflexClientHandlingSystem
         private void schHome_Click(object sender, EventArgs e)
         {
             ProjectManager pm = new ProjectManager();
+            this.Hide();
+            pm.ShowDialog();
             this.Close();
-            pm.Show();
         }
 
         private void addEvent_Click(object sender, EventArgs e)
@@ -102,7 +113,7 @@ namespace ResoflexClientHandlingSystem
 
             string projName = projectName.Text.ToString();
 
-            string sql = "select s.sch_no as Schedule_No, s.proj_id, s.visit_type_id, p.proj_name as Project_Name, vt.type as Schedule_Type, s.from_date_time as Start_Date_and_Time, s.to_date_time as End_Date_and_Time, s.to_do_list as TODO_List, s.resource as Resources, s.check_list as Check_List, s.travelling_mode as Travelling_Mode, s.accommodation as Accomodation, s.meals as Meals, s.logs as Logs " +
+            string sql = "select s.sch_no as Schedule_No, s.proj_id, s.visit_type_id, p.proj_name as Project_Name, vt.type as Schedule_Type, s.from_date_time as Start_Date_and_Time, s.to_date_time as End_Date_and_Time, s.to_do_list as TODO_List, s.check_list as Check_List, s.travelling_mode as Travelling_Mode, s.accommodation as Accomodation, s.meals as Meals, s.logs as Logs " +
                          " from schedule s, project p, visit_type vt " +
                          " where(s.proj_id = p.proj_id) and (s.visit_type_id = vt.visit_type_id) and (p.proj_name like '%" + projName + "%') " +
                          " order by s.sch_no, s.proj_id;";
@@ -111,19 +122,11 @@ namespace ResoflexClientHandlingSystem
             {
                 MySqlDataReader reader = DBConnection.getData(sql);
 
-                if (reader.HasRows)
-                {
-                    DataTable dt = new DataTable();
-                    dt.Load(reader);
+                DataTable dt = new DataTable();
+                dt.Load(reader);
+                scheduleGrid.DataSource = dt;
 
-                    scheduleGrid.DataSource = dt;
-                }
-                else
-                {
-                    //scheduleGrid.DataSource = null;
-
-                    reader.Close();
-                }
+                reader.Close();
             }
             catch (Exception)
             {
@@ -138,28 +141,20 @@ namespace ResoflexClientHandlingSystem
 
             string cName = clientName.Text.ToString();
 
-            string sql = "select s.sch_no as Schedule_No, s.proj_id, s.visit_type_id, p.proj_name as Project_Name, vt.type as Schedule_Type, s.from_date_time as Start_Date_and_Time, s.to_date_time as End_Date_and_Time, s.to_do_list as TODO_List, s.resource as Resources, s.check_list as Check_List, s.travelling_mode as Travelling_Mode, s.accommodation as Accomodation, s.meals as Meals, s.logs as Logs" +
+            string sql = "select s.sch_no as Schedule_No, s.proj_id, s.visit_type_id, p.proj_name as Project_Name, vt.type as Schedule_Type, s.from_date_time as Start_Date_and_Time, s.to_date_time as End_Date_and_Time, s.to_do_list as TODO_List, s.check_list as Check_List, s.travelling_mode as Travelling_Mode, s.accommodation as Accomodation, s.meals as Meals, s.logs as Logs" +
                          " from schedule s, project p, visit_type vt, client c " +
-                         " where (s.proj_id = p.proj_id) and (s.visit_type_id = vt.visit_type_id) and (p.client_id = c.client_id) and (c.name like '%"+ cName + "%') " +
+                         " where (s.proj_id = p.proj_id) and (s.visit_type_id = vt.visit_type_id) and (p.client_id = c.client_id) and (c.name like '%" + cName + "%') " +
                          " order by s.sch_no, s.proj_id;";
 
             try
             {
                 MySqlDataReader reader = DBConnection.getData(sql);
 
-                if (reader.HasRows)
-                { 
-                    DataTable dt = new DataTable();
-                    dt.Load(reader);
+                DataTable dt = new DataTable();
+                dt.Load(reader);
+                scheduleGrid.DataSource = dt;
 
-                    scheduleGrid.DataSource = dt;
-                }
-                else
-                {
-                    //scheduleGrid.DataSource = null;
-
-                    reader.Close();
-                }
+                reader.Close();
             }
             catch (Exception)
             {
@@ -173,6 +168,7 @@ namespace ResoflexClientHandlingSystem
         {
             Schedule sch = new Schedule();
             ArrayList serviceEng = new ArrayList();
+            ArrayList resoArray = new ArrayList();
 
             MySqlDataReader reader1 = DBConnection.getData("select * from schedule where sch_no =" + schNo + " and proj_id = " + proj_id + ";");
 
@@ -201,15 +197,29 @@ namespace ResoflexClientHandlingSystem
                     staff.StaffId = int.Parse(reader2.GetString("staff_id"));
                     staff.FirstName = reader2.GetString("first_name");
                     staff.LastName = reader2.GetString("last_name");
-
-                    //MessageBox.Show(int.Parse(reader2.GetString("staff_id")) + reader2.GetString("first_name") + reader2.GetString("last_name"));
-
+                    
                     serviceEng.Add(staff);
                 }
 
                 sch.ServEngineer = serviceEng;
 
                 reader2.Close();
+
+                MySqlDataReader reader3 = DBConnection.getData("select sr.resource_id, sr.qty, r.name from schedule_resources sr, resource r where ( sr.sch_no =" + schNo + " and sr.proj_id = " + proj_id + ") and (sr.resource_id = r.resource_id);");
+
+                while (reader3.Read())
+                {
+                    Resource reso = new Resource();
+                    reso.ResourceId = int.Parse(reader3.GetString("resource_id"));
+                    reso.Name = reader3.GetString("name");
+                    reso.TotalQty = int.Parse(reader3.GetString("qty"));
+
+                    resoArray.Add(reso);
+                }
+
+                sch.ResoArray = resoArray;
+
+                reader3.Close();
 
             }
             else
@@ -238,11 +248,11 @@ namespace ResoflexClientHandlingSystem
         }
 
         //Data for grid
-        private static DataTable getSchedules()
+        private DataTable getSchedules()
         {
             DataTable dt = new DataTable();
 
-            MySqlDataReader reader = DBConnection.getData("select s.sch_no as Schedule_No, s.proj_id, s.visit_type_id, p.proj_name as Project_Name, vt.type as Schedule_Type, s.from_date_time as Start_Date_and_Time, s.to_date_time as End_Date_and_Time, s.to_do_list as TODO_List, s.resource as Resources, s.check_list as Check_List, s.travelling_mode as Travelling_Mode, s.accommodation as Accomodation, s.meals as Meals, s.logs as Logs " +
+            MySqlDataReader reader = DBConnection.getData("select s.sch_no as Schedule_No, s.proj_id, s.visit_type_id, p.proj_name as Project_Name, vt.type as Schedule_Type, s.from_date_time as Start_Date_and_Time, s.to_date_time as End_Date_and_Time, s.to_do_list as TODO_List, s.check_list as Check_List, s.travelling_mode as Travelling_Mode, s.accommodation as Accomodation, s.meals as Meals, s.logs as Logs " +
                                                             "from schedule s, project p, visit_type vt " +
                                                             "where (s.proj_id = p.proj_id) and (s.visit_type_id = vt.visit_type_id) " +
                                                             " order by s.sch_no, s.proj_id;");
@@ -250,6 +260,9 @@ namespace ResoflexClientHandlingSystem
             dt.Load(reader);
 
             reader.Close();
+
+            totalScheduleTile();
+            incompleteScheduleTile();
 
             return dt;
         }
@@ -298,6 +311,59 @@ namespace ResoflexClientHandlingSystem
         private void btnRefresh_Click(object sender, EventArgs e)
         {
             scheduleGrid.DataSource = getSchedules();
+        }
+
+        //data for tiles
+        public void totalScheduleTile()
+        {
+            MySqlDataReader reader = DBConnection.getData("select count(*) as count from schedule;");
+
+            if (reader.Read())
+            {
+                totalScheduels.Text = reader.GetInt16("count").ToString();
+            }
+            else
+            {
+                totalScheduels.Text = 0.ToString();
+            }
+
+            reader.Close();
+        }
+
+        public void incompleteScheduleTile()
+        {
+            int count = 0;
+
+            MySqlDataReader reader = DBConnection.getData("select count(*) as count from schedule s, event e where s.sch_no = e.sch_no and s.proj_id = e.proj_id and e.to_date_time > NOW()");
+
+            if (reader.Read())
+            {
+                count = reader.GetInt16("count");   
+            }
+            
+            reader.Close();
+
+            MySqlDataReader reader1 = DBConnection.getData("select count(*) as count from schedule s, event e where s.sch_no != e.sch_no or s.proj_id != e.proj_id");
+
+            if (reader1.Read())
+            {
+                count += reader1.GetInt16("count");
+            }
+            
+            reader1.Close();
+
+            incompleteSchedules.Text = count.ToString();
+        }
+
+        private void reports_Click(object sender, EventArgs e)
+        {
+            DataGridViewRow row = scheduleGrid.CurrentRow;
+
+            int sch_no = int.Parse(row.Cells[0].Value.ToString());
+            int proj_id = int.Parse(row.Cells[1].Value.ToString());
+
+            ScheduleReportsForm srf = new ScheduleReportsForm(proj_id, sch_no);
+            srf.ShowDialog();
         }
     }
 }
